requirements.txt
pycryptodome>=3.20.0
pydantic>=2.5.3
typer>=0.9.0
rich>=13.7.0
pyperclip>=1.8.2
bcrypt>=4.0.1
